Source code added
The whole program source code is in the single file gameOfLife.c. The code compiles and runs as intended on Linux and Windows.

The current version of the program is not interactive. To change behaviour edit the (minimalistic) main function. You can reuse the code from the demo functions randomGameDemo and gliderGunDemo for this purpose, or alter these functions themselves.